+++ src/mopidy_podcast_itunes/__init__.py
@@ -4,8 +4,6 @@ from mopidy import config, ext, httpclient
 
 __version__ = "3.1.0"
 
-CHARTS = ["podcasts", "audioPodcasts", "videoPodcasts"]
-
 COUNTRIES = [
     "AD",
     "AE",
@@ -278,12 +276,12 @@ class Extension(ext.Extension):
             base_url=config.String(),
             country=config.String(choices=COUNTRIES),
             explicit=config.String(choices=EXPLICIT, optional=True),
-            charts=config.String(choices=CHARTS),
             charts_limit=config.Integer(minimum=1, maximum=MAX_LIMIT, optional=True),
             search_limit=config.Integer(minimum=1, maximum=MAX_LIMIT, optional=True),
             timeout=config.Integer(minimum=1, optional=True),
             retries=config.Integer(minimum=0),
             # no longer used
+            charts=config.Deprecated(),
             charts_format=config.Deprecated(),
             episode_format=config.Deprecated(),
             genre_format=config.Deprecated(),